Function zelkova_lang::compiler::check_module

pub fn check_module(
    package: &PackageName, 
    interfaces: &HashMap<Name, Interface>, 
    source: &Module
) -> Result<Module, CompilationError>

Take a parsed module file within the ecosystem and apply all checks to it

TODO canonicalization must happens before checkings, because type check (at least) will require access to other modules canonical representation. That probably mean moving the canonical::canonicalize call out of this function