zelkova_lang::compiler::parser

Module layout

Simplify the indentation manager for the parser by doing it in before the token iterator is passed to the parser.

Structs

Offside

Enums

Context

Context represent the kind of expression we are looking at.

LayoutError

Functions

layout