Module zelkova_lang::compiler::parser::tokenizer

Module providing a transformation from a textual source code to a serie of tokens.

Directly inspired by the great work on the RustPython team https://github.com/RustPython/RustPython/blob/master/parser/src/lexer.rs

Structs

TokenizerError
Represents an error during tokenization.

Enums

Token
Represents the different part which constitute our source code

TokenizerErrorType
The type of error refered in TokenizerError

Functions

make_tokenizer
Take a source code and return an iterator of Spanned

Type Definitions

Result
Represent a standard Result scoped to a TokenizerError