Module zelkova_lang::compiler::typer

This module contains the type checker pass of the language

It works with the source AST and will perform two jobs:

type checks the different declarations and expression

infer the types when not declared in the source

I have no idea how that works; so bear with me while I explore the space, make mistake and (hopefully) learn something :)

Some papers on type inference:

http://steshaw.org/hm/hindley-milner.pdf

https://pdfs.semanticscholar.org/8983/233b3dff2c5b94efb31235f62bddc22dc899.pdf

http://gallium.inria.fr/~fpottier/publis/fpottier-elaboration.pdf

http://gallium.inria.fr/~fpottier/publis/emlti-final.pdf

A type inference problem consists of a type environment Γ , an expression t, and a type T of kind ?

Constraint generation rules:

Equation 1: ⟦x : T⟧ = x ≼ T “x has type T if and only if T is an instance of the type scheme associated with x” Important part: There is no relation to the typing environment Γ, instead x appears free (and will be bound to Γ later)

Equation 2: ⟦λz.t : T⟧ = ∃X1X2.(let z : X1 in ⟦t : X2⟧ ∧ X1 → X2 ≤ T) “λz.t has type T if and only if, for some X1 and X2, (i) under the assumption that z has type X1, t has type X2, and (ii) T is a supertype of X1 → X2.” z and t types must be fresh (can’t generally guess them). They are existentially bound because we are going to solve their values. Note that z is not fresh in the condition (i).

Equation 3: ⟦t1 t2 : T⟧ = ∃X2.(⟦t1 : X2 → T⟧ ∧ ⟦t2 : X2⟧) “t1 t2 has type T if and only if, for some X2, t1 has type X2 → T and t2 has type X2”

Equation 4: ⟦let z = t1 in t2 : T⟧ = let z : ∀X[⟦t1 : X⟧].X in ⟦t2 : T⟧ “let z = t1 in t2 has type T if and only if, under the assumption that z has every type X such that ⟦t1 : X⟧ holds, t2 has type T”

Structs

TypeVariable

Enums

Error

Term

untyped term. In zelkova that would be the parsed source (or canonical, not sure yet)

Type

TypeLiteral

Functions

infer

infer the type of the given term given known function defined in the outer scopes. This is a translation of the algorithm demonstrated by Ionut Gan at I T.A.K.E Unconference 2015

type_check