Module zelkova_lang::compiler

The Zelkova compiler

How to compile a package ?

Note: We don’t manage interfaces and external modules. We try to keep them in mind, so they can be relatively easily added later on.

Start at the src/ folder. We name it root. (later on defined in a zelkova.json manifest)

Collect all *.zelkova files with their path name relatives to the root.

Create a SourceFiles mapping from ModuleName to parser::Module.

module names are deduced from file name

parsing is done through parser::parse

Verify that parser::Module.name match the one from the file system

Build a dependency graphs from the modules import

build it

Verify there is no cyclic relation between modules

Following the deps graph,

canonicalize each modules

check each module (type check, exhaustiveness, etc…)

Bonus point to parallelize the tree branches which are not dependent on each others

Once we have a module with all checks passing, create its interface and emit AST/interface

TODO Try a way to weave error management in each of those passes :)

Modules

canonical

The canonical representation of a zelkova programs is a translation of a local source into the broader world.

name

Home to the Name and QualName primitives

parser

This module contains all the data types representing the source of our language.

position

source

typer

This module contains the type checker pass of the language

Structs

Interface

An interface is trim down version of a module.

ModuleName

A module name represent

PackageName

A package name is composed of an author and project name and is written as author/project.

Enums

CompilationError

Functions

check_module

Take a parsed module file within the ecosystem and apply all checks to it

compile_package