Module zelkova_lang::compiler::canonical

The canonical representation of a zelkova programs is a translation of a local source into the broader world.

This phase is where we integrate the local parsed module into the rest of the program. We do the following steps:

Resolve all imports

Qualify all Name (eg. a local value test in a Mod.A module will be renamed Mod.A.test)

Checks that exported names are actually present in the module

Checks there is none cyclic dependency between this module and others (Might be done earlier, let’s see)

Note that we use HashMap’s a lot in this module’s structures. This is because later phases will want to have cheap access to the different components of a Module.

TODO Rename this to core ? I feel it’s going to be te main internal representation of the language.

Re-exports

pub use parser::Associativity;

Structs

CaseBranch

Infix

Module
A resolved module

TypeConstructor

UnionType

Enums

Error

ExportType

Exports

Expression
Expression is an optimized version for checks and caches.

Pattern

Type

Value

Functions

canonicalize
Transform a given parser::Module into a canonical::Module