Module zelkova_lang::compiler::parser

This module contains all the data types representing the source of our language.

This is the part the user interact with, and contains some syntax sugar to make their life easier (for example, pattern matching in function declaration)

TODO Fill in the sections below

Compiler phase

parse

AST

TODO

Modules

TODO hierarchy and modules.

Re-exports

pub use error::Error;

Modules

error

layout

Simplify the indentation manager for the parser by doing it in before the token iterator is passed to the parser.

tokenizer

Module providing a transformation from a textual source code to a serie of tokens.

Structs

CaseBranch

FunBinding

A FunBinding is one of the (possibly multiple) function declaration.

FunType

Represents the type signature of a particular function

Function

Function represent a function declaration in the source code.

Import

A representation of the import declaration

Infix

Match

The match structure is composed of a serie of patterns and an associated expression

Module

A Module is the top-level structure for a source file.

UnionType

Enums

Associativity

Declaration

A Declaration is a top-level block and is the basis for a Module.

Exposed

Exposed represent the terms exposed/imported by a module.

Exposing

Exposing represent whether an import (or export) expose terms.

Expression

An Expression

Literal

A literal

Pattern

A pattern is the left handside of a pattern-match expression

Privacy

Privacy control how a custom type is exposed.

Type

A part of a declared type. This is also used in type annotations.

Functions

parse