zelkova_lang::compiler::parser::tokenizer

Function make_tokenizer

pub fn make_tokenizer<'a>(
    source: &'a str,
) -> impl Iterator<Item = Result<Spanned<Position, Token>>> + 'a

Take a source code and return an iterator of Spanned

This is the access point of this module, and the only way to create the underlying Tokenizer (although it isn’t exposed to public consumption)